+++ sympy/core/expr.py
@@ -4103,9 +4103,9 @@ def _mag(x):
     if not xpos:
         return S.Zero
     try:
-        mag_first_dig = int(ceil(log10(xpos)))
+        mag_first_dig = ceil(log10(xpos))
     except (ValueError, OverflowError):
-        mag_first_dig = int(ceil(Float(mpf_log(xpos._mpf_, 53))/log(10)))
+        mag_first_dig = ceil(Float(mpf_log(xpos._mpf_, 53))/log(10))
     # check that we aren't off by 1
     if (xpos/S(10)**mag_first_dig) >= 1:
         assert 1 <= (xpos/S(10)**mag_first_dig) < 10
