+++ sympy/simplify/hyperexpand.py
@@ -2320,7 +2320,7 @@ def _meijergexpand(func, z0, allow_hyper=False, rewrite='default',
                 integrand = z**s
                 for b in bm:
                     if not Mod(b, 1) and b.is_Number:
-                        b = int(round(b))
+                        b = round(b)
                     integrand *= gamma(b - s)
                 for a in an:
                     integrand *= gamma(1 - a + s)
@@ -2332,7 +2332,7 @@ def _meijergexpand(func, z0, allow_hyper=False, rewrite='default',
                 # Now sum the finitely many residues:
                 # XXX This speeds up some cases - is it a good idea?
                 integrand = expand_func(integrand)
-                for r in range(int(round(lu))):
+                for r in range(round(lu)):
                     resid = residue(integrand, s, b_ + r)
                     resid = apply_operators(resid, ops, lambda f: z*f.diff(z))
                     res -= resid
